+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Unreleased
+
+- **Other Features**
+    - Added `Spiral\Debug\Config\DebugConfig` has been added for easy addition of tags and collectors.
+
 ## 3.10.0 - 2023-11-24
 
 - **Other Features**